Method setCursor: change type of array row_offsets.

+++ Deuligne.cpp
@@ -126,7 +126,7 @@ void Deuligne::begin(uint8_t cols, uint8_t lines, uint8_t dotsize) {
 }
 
 void Deuligne::setCursor(uint8_t col, uint8_t row) {
-  int row_offsets[] = { 0x00, 0x40, 0x14, 0x54 };
+  uint8_t row_offsets[] = { 0x00, 0x40, 0x14, 0x54 };
   if ( row > myNumLines ) {
     row = myNumLines-1;    // we count rows starting w/0
   }
